There is now a distinction between a data type which indicates the converter used for serialization and a wire type which indicates what actual data is written on the wire. This type is used to select the converter for deserialization.
As the registry concept from RSC is not portable to windows without flaws, converters must now be registered explicitly everytime in a new ConverterCollection class. For default converters this is automatically done in a new RSBFactory class which now forms the entry point for the client interface. This is a manual merge from the ExternalSpread branch.