Prevent the presence of multiple singleton factories on windows.
Due to the separate memory management of windows for different DLLs the
existing singleton implementation does not work there. All uses of the
singleton were replace by appropriate getters to a specific instance.

In detail:
* Singleton base classes were made private to reuse the instantiation
support in static initializer code of the existing singleton
implementation
* Appropriate getter methods were added falling back to the now private
singleton implementation
* getInstance methods were redeclared using custom code going back to
the getters. These methods are now marked deprecated and singleton use
is discouraged.

For EventSendingStrategyFactory and the receiving counterpart these
changes could not be performed because they use a specific factory
implementation from RSC. However, they are probably never used from the
outside.

refs #1245