Fix Windows compilation
Commit 3d68ebe6e broke the windows build by changing the name of the
target built by CMake. This, in turn, changed the compile time
definition CMake generates to set the dllexport flags for Windows. As
this define name change was not reflected in rsbexports.h.in, exporting
was basically broken since then.

This commit reflects the change. As the new target name contains a .
character, we have to manually derive the name of the compile time
definition that is generated by CMake to exclude the dot character
before generating the header file.