Do not abort on failed shutdown in src/rsb/transport/socket/BusConnection.cpp
fixes #1846

Apparently, an exception can be thrown during socket shutdown when the
remote peer requests a shutdown and the local side is not quick enough
with its shutdown.

This seems like incorrect behavior of the remote peer but has to be
handled anyway since e.g. a crash of the remote program would have an
identical effect.

* src/rsb/transport/socket/BusConnection.cpp (header): updated copyright
  (BusConnection::handleReadLength): catch exceptions thrown by
  shutdown, print a warning but ignore otherwise