Add parallel call parameter to ParallelEventReceivingStrategy
Adds a config option to ParallelEventReceivingStrategy that allows to
specify whether individual handlers might be called in parallel with
potentially overlapping calls. This might result in situations that
conflict with the specified QoS-settings because if the new parameter is
set to true, dispatched events might appear out of order to the client
code due to thread scheduling issues.

refs #1870