Correctly handle dll export symbols on windows
The introspection code is a client of the main RSB library and therefore
needs to dllimport the protocol symbols instead of exporting it. In
order to implement this change major parts of the existing symbol
declaration logic have been converted to target properties which act
more locally than the general add_definitions calls that existed before.