Inprocess transp. independent of static destruction
This commit makes the inprocess transport independent of static
destruction order so that participants using this transport can easily
be used inside static variables without errors. The logic for this
implementation is to couple the lifetime of the whole transport to the
lifetime of participants by consequently using shared pointers for this
purpose.

* src/rsb/transport/inprocess/Bus.{cpp,h}
  Remove singleton implementation to prevent static destruction without
  shared ptrs leaving scope. Introduce a method to retrieve a default
  instance that is maintained as a shared ptr instead.
* src/rsb/transport/inprocess/InConnector.{cpp,h}
  Add the Bus instance this connector is operating on as a shared ptr so
  that the connector keeps up the lifetime of the bus instance
* src/rsb/transport/inprocess/OutConnector.{cpp,h}
  likewise

refs #2034