Reset factoryWhileLoadingPlugins in src/rsb/Factory.cpp
fixes #2528, refs #2487

The Factory constructor could be unwound by an exception thrown by the
plugin system, leaving factoryWhileLoadingPlugins pointing to a
half-initialized Factory object. To prevent this, reset
factoryWhileLoadingPlugins to NULL in case an exception is thrown by the
plugin system, then re-throw.

Unfortunately, I cannot come up with a reasonable test for this. The
error would be easy to trigger by doing

  setenv("RSB_PLUGINS_CPP_LOAD", "foo", 1);
  try { Factory f; } catch (...) {};
  Factory f => segmentation fault

but that would require making the Factory constructor public or some
equally ugly friend hackery. *sigh*

* src/rsb/Factory.cpp (header): updated copyright
  (Factory::Factory): reset factoryWhileLoadingPlugins in case an
  exception is thrown